Add example datasets and Newick tree popup window.
This commit adds Julie's examples to the interface and adds a link to
view the Newick string at the end of the workflow.

The string contains NHX IMG tags, which provides thumbnail images for
species. Compatible tree viewers (erm, at least phylowidget) will show
these as images in the tree.